// cypress/cypress/support/helpers/enums/btn-names-enums.ts
/*this enum created for different name the submit button
for login, OTP, registration, profile and lobby pages
 */

export enum submitBtn {
  submit = 'Submit',
  continue = 'Continue',
  register = 'Register',
  update = 'Update',
  change = 'Change',
  login = 'Login',
  sendOTP = 'Send OTP',
  confirm = 'Confirm',
  review = 'Review'
}

export enum commonBtn {
  sortBy = 'Sort by...',
  favs = 'Favorites',
  moreGame = 'More Games',
  moreJackpot = 'More Jackpots'
}

export enum allOtherPageBtns {
  cookieBtn = 'I agree', // cookie notice button
  reload = 'Reload Page', // 500 error page
  home = 'Home Page', // 404 error page
  resend = 'Resend code', //
  loadMore = 'Load more'
}

export enum backLobbyBtn {
  liveCasino = 'Live Casino',
  liveSlots = 'Live Slots',
  eGames = 'e-Games'
}

export enum slotsBtnsTitles {
  classicView = 'Turn on Classic View',
  modernView = 'Turn on Modern View',
  jackpots = 'Jackpots',
  provider = 'Provider',
  vacancy = 'Vacancy',
  standalone = 'Standalone'
}

export enum eGamesBtns {
  play = 'Play Now',
  demo = 'Demo Play',
  cancel = 'Cancel',
  try = 'Try It Now!'
}

export enum bannerBtns {
  playLive = 'Play Live',
  rewardsSolaire = 'Solaire Rewards',
  bookNow = 'Book now'
}

export enum depositBtns {
  amount5000 = '5,000',
  amount10000 = '10,000',
  amount25000 = '25,000',
  back = 'Back'
}